+++ nautilus_trader/model/enums.pyx
@@ -32,12 +32,11 @@ from nautilus_trader.core.rust.model import OptionKind
 from nautilus_trader.core.rust.model import OrderSide
 from nautilus_trader.core.rust.model import OrderStatus
 from nautilus_trader.core.rust.model import OrderType
+from nautilus_trader.core.rust.model import PositionSide
 from nautilus_trader.model.c_enums.instrument_close_type import InstrumentCloseType
 from nautilus_trader.model.c_enums.instrument_close_type import InstrumentCloseTypeParser
 from nautilus_trader.model.c_enums.instrument_status import InstrumentStatus
 from nautilus_trader.model.c_enums.instrument_status import InstrumentStatusParser
-from nautilus_trader.model.c_enums.position_side import PositionSide
-from nautilus_trader.model.c_enums.position_side import PositionSideParser
 from nautilus_trader.model.c_enums.price_type import PriceType
 from nautilus_trader.model.c_enums.price_type import PriceTypeParser
 from nautilus_trader.model.c_enums.time_in_force import TimeInForce
@@ -85,6 +84,8 @@ from nautilus_trader.core.rust.enums cimport order_status_from_str
 from nautilus_trader.core.rust.enums cimport order_status_to_str
 from nautilus_trader.core.rust.enums cimport order_type_from_str
 from nautilus_trader.core.rust.enums cimport order_type_to_str
+from nautilus_trader.core.rust.enums cimport position_side_from_str
+from nautilus_trader.core.rust.enums cimport position_side_to_str
 
 
 __all__ = [
@@ -110,7 +111,6 @@ __all__ = [
     "OrderStatus",
     "OrderType",
     "PositionSide",
-    "PositionSideParser",
     "PriceType",
     "PriceTypeParser",
     "TimeInForce",
@@ -157,4 +157,6 @@ __all__ = [
     "order_status_from_str",
     "order_type_to_str",
     "order_type_from_str",
+    "position_side_to_str",
+    "position_side_from_str",
 ]