+++ nautilus_trader/model/events/position.pyx
@@ -20,11 +20,12 @@ from libc.stdint cimport uint64_t
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.message cimport Event
 from nautilus_trader.core.rust.enums cimport OrderSide
+from nautilus_trader.core.rust.enums cimport PositionSide
 from nautilus_trader.core.rust.enums cimport order_side_from_str
 from nautilus_trader.core.rust.enums cimport order_side_to_str
+from nautilus_trader.core.rust.enums cimport position_side_from_str
+from nautilus_trader.core.rust.enums cimport position_side_to_str
 from nautilus_trader.core.uuid cimport UUID4
-from nautilus_trader.model.c_enums.position_side cimport PositionSide
-from nautilus_trader.model.c_enums.position_side cimport PositionSideParser
 from nautilus_trader.model.currency cimport Currency
 from nautilus_trader.model.events.order cimport OrderFilled
 from nautilus_trader.model.identifiers cimport AccountId
@@ -164,7 +165,7 @@ cdef class PositionEvent(Event):
             f"opening_order_id={self.opening_order_id.to_str()}, "
             f"closing_order_id={self.closing_order_id}, "  # Can be None
             f"entry={order_side_to_str(self.entry)}, "
-            f"side={PositionSideParser.to_str(self.side)}, "
+            f"side={position_side_to_str(self.side)}, "
             f"net_qty={self.net_qty}, "
             f"quantity={self.quantity.to_str()}, "
             f"peak_qty={self.peak_qty.to_str()}, "
@@ -191,7 +192,7 @@ cdef class PositionEvent(Event):
             f"opening_order_id={self.opening_order_id.to_str()}, "
             f"closing_order_id={self.closing_order_id}, "  # Can be None
             f"entry={order_side_to_str(self.entry)}, "
-            f"side={PositionSideParser.to_str(self.side)}, "
+            f"side={position_side_to_str(self.side)}, "
             f"net_qty={self.net_qty}, "
             f"quantity={self.quantity.to_str()}, "
             f"peak_qty={self.peak_qty.to_str()}, "
@@ -353,7 +354,7 @@ cdef class PositionOpened(PositionEvent):
             account_id=AccountId(values["account_id"]),
             opening_order_id=ClientOrderId(values["opening_order_id"]),
             entry=order_side_from_str(values["entry"]),
-            side=PositionSideParser.from_str(values["side"]),
+            side=position_side_from_str(values["side"]),
             net_qty=values["net_qty"],
             quantity=Quantity.from_str_c(values["quantity"]),
             peak_qty=Quantity.from_str_c(values["peak_qty"]),
@@ -379,7 +380,7 @@ cdef class PositionOpened(PositionEvent):
             "account_id": obj.account_id.to_str(),
             "opening_order_id": obj.opening_order_id.to_str(),
             "entry": order_side_to_str(obj.entry),
-            "side": PositionSideParser.to_str(obj.side),
+            "side": position_side_to_str(obj.side),
             "net_qty": obj.net_qty,
             "quantity": str(obj.quantity),
             "peak_qty": str(obj.peak_qty),
@@ -612,7 +613,7 @@ cdef class PositionChanged(PositionEvent):
             account_id=AccountId(values["account_id"]),
             opening_order_id=ClientOrderId(values["opening_order_id"]),
             entry=order_side_from_str(values["entry"]),
-            side=PositionSideParser.from_str(values["side"]),
+            side=position_side_from_str(values["side"]),
             net_qty=values["net_qty"],
             quantity=Quantity.from_str_c(values["quantity"]),
             peak_qty=Quantity.from_str_c(values["peak_qty"]),
@@ -642,7 +643,7 @@ cdef class PositionChanged(PositionEvent):
             "account_id": obj.account_id.to_str(),
             "opening_order_id": obj.opening_order_id.to_str(),
             "entry": order_side_to_str(obj.entry),
-            "side": PositionSideParser.to_str(obj.side),
+            "side": position_side_to_str(obj.side),
             "net_qty": obj.net_qty,
             "quantity": str(obj.quantity),
             "peak_qty": str(obj.peak_qty),
@@ -883,7 +884,7 @@ cdef class PositionClosed(PositionEvent):
             opening_order_id=ClientOrderId(values["opening_order_id"]),
             closing_order_id=ClientOrderId(values["closing_order_id"]),
             entry=order_side_from_str(values["entry"]),
-            side=PositionSideParser.from_str(values["side"]),
+            side=position_side_from_str(values["side"]),
             net_qty=values["net_qty"],
             quantity=Quantity.from_str_c(values["quantity"]),
             peak_qty=Quantity.from_str_c(values["peak_qty"]),
@@ -914,7 +915,7 @@ cdef class PositionClosed(PositionEvent):
             "opening_order_id": obj.opening_order_id.to_str(),
             "closing_order_id": obj.closing_order_id.to_str(),
             "entry": order_side_to_str(obj.entry),
-            "side": PositionSideParser.to_str(obj.side),
+            "side": position_side_to_str(obj.side),
             "net_qty": obj.net_qty,
             "quantity": str(obj.quantity),
             "peak_qty": str(obj.peak_qty),